fix jquery live events on popoverview
The live click event on "add property", "cancel" and "create new
resource/save resource" was called a bunch of times if the popoverview is
shown without reload workaround. Before set a live click event, all
other live click events will be unbinded. So it is possible to click on
cancel without reload the whole page.