first test to submit values without reloading
This is a first test to see if it is possible to submit value
with RDFauthor without reloading the whole site each time.
There are many known issues that have to be treated later. Also it will
right now only work for literal and resource widgets. Also the reloading
has to be disabled in the support.js (silverblue theme), which is done in a
separate commit.