Fix: workaround for bug when working with xsd:date
This is an ad-hoc fix for a problem that is somewhat
hard to trace ('Uncaught exceptions [object Object]') that occurs
specifically when an empty string with datatype 'xsd:date' is passed
to jQuery.rdf.literal(). This is actually a pathologic case (the empty
string is not a date), but can occur when creating new intances and or
adding new propertiesi to a resource.

The default behaviour of RDFauthor is to never carry over datatypes
when the value is empty/not set. This will be changed because due to
this behaviour the wrong (i.e. default) widget is selected for newly
created dates/times. Note that this change will happen the OntoWiki
repo due to the nature of RDFauthor.