Widget for literals do not create empty literals
When calling 'Create Instance', there are properties that might be
give a value or not. Since we changed the RDFauthor in such a way, that
empty value aren't handled as null anymore, all properties will be created
with empty values, when creating a new resource. This might or might not
be wanted, so we pass the working mode to the widgets and handle empty
literals differently based on the working mode.