Fix labels in RDFauthor-Popoverview
When data is passed to the RDFauthor via the rdfauthorinit-action of the
service controller, a title as also passed. Internally, the RDFauthor uses
statements defined in rdfauthor.statement.js as well as triples of the
jquery.rdf library. When converting back and forth between statements and
triples, the label gets lost (because a rdfauthor.statement has more
information than a query.rdf triples, specifically there is a key
'_predicateLabel' in the rdfauthor.statement that is not carried over).
As a quick solution, we extend the predicate of the jquery.triple by adding
a key 'label' to it.