Test to enable drop down selections for RDFauthor
This is a test to enable dropdown selections for the RDFauthor. The new
dropdown widget is a modification of the literal widget which needs to be
customized further. There is also only minimal error handling right now.

Usage: If a (datatype) predicate has an rdfs:range A and A has an owl:oneOf,
the dropdown widget will be used, i.e. in this situation:

    :predicate a owl:DatatypeProperty .
    :predicate rdfs:range :dropDownExample .
    :dropDownExample owl:oneOf [ "Ex1"^^xsd:string, "Ex2"^^xsd:string ] .