copyWithObject will respect a passed type
When editing with RDFauthor, there is a method called
copyWithObject that copies a statement's subject and predicate, but
exchanges the old object for a new one. However, there is a line of code
that explictly adds the type of the old object to the new object.
We now check first if the new object has a type and only add the old type
if no type is given.