Handle optional propertyOrder
The feature sortPropertiesForRDFauthor provides a sorting order for the
predicates. This commit will handle this optional argument.

NB: There will be some refactoring and the sorting order will be passed
differenly in the near future. So this will be changed soon. Also there
will be a separate branch that enables this feature without the form builder
changes.