Add i18n options to dropdowns
This is a first draft on how to provide an i18n for the RDFauthor. We use
a global variable RDFAUTHOR_LANGUAGE that holds the language as string.
To work with OntoWiki, this variable can be injected in the global scope
via OntoWiki/application/classes/OntoWiki/Controller/Base.php, i.e.

'var RDFAUTHOR_LANGUAGE = "' . $this->_config->languages->locale . '";'

as part of the string jsonVars. NB: The line above can be copies verbatim
WITHOUT the enclosing single quotes into line 129 of the Base.php