Slight refactoring concerning commit w/o reload
The RDFauthor has recently been factored to update the changed values
in the view without reloading the whole page. In order to always display
values that are actually in the triple store (as opposed to 'value that have
just been entered'), the values are queried again right after storing them
(might be slightly paranoid). With this commit, the
function to query the stored values is moved into the callback of the ajax
request that commits the values.