Removed dropdown widgets return value null
When a widget gets removed (via the 'minus' button) it is in fact still
active and hidden. When submitting a change, RDFauthor tries to fetch a
value even from the removed widgets. Usually they return null, if there is
no value. In case of the dropdown widgets, there was an error which is fixed
in this commit.