SPARQL10, SPARQL11 and exhaustive methods
This commit refactures the dropdown widget somewhat. In the rdfauthor.config.js,
the dropdown gets two additional keys, namely 'method' and 'use_exhaustive'.
The value for 'method' can either be 'sparql10' or 'sparql11', the value
for 'use_exhautive' is a boolean value.

'sparql10' will use a (rather slow) SPARQL 1.0 query which can only query
collections/dropdowns up to a fixed length. This is the default value, if
no value was found.
'sparql11' will use a SPARQL 1.1 query which is more flexible and can query
collections/dropdowns of arbitrary length and is also faster (but the query
uses property paths which are not supported by the vanilla OntoWiki/Erfurt).

The triples that (collectively) trigger a dropdown widget for ?PROPERTY are:

    ?PROPERTY rdfs:range ?CLASS .
    ?CLASS owl:oneOf ( ?RES1 ?RES2 ?RES3 ... ) .

'use_exhaustive' will (independently of the method) query all values of the
range class ?RANGE. To trigger this behavious, there must be triples of the
form:

    ?PROPERTY rdfs:range ?RANGE .
    ?RANGE sysOnt:displayAs "dropdown" .
    ?RES1 a ?RANGE .
    ?RES2 a ?RANGE .
    ...

Note: 'use_exhaustive' is a quick method and it also circumvents the problems
OntoWiki has with blank nodes.

There might be a better way to express the 'closedness' of a class using owl
constructs, so this might be changed. If the key 'use_exhaustive' is set to
false or not set at all, this behaviour will never be used.