Add year dropdown for calendar widget
When xsd:date (not datetime or time) is used, the calendar widget will
also show a year.

NB: The default.css of the silverblue theme messes the dropdown up, the CSS
should reset the min-width like so (somewhere round line 1300):

select.ui-datepicker-year {
    min-width: initial;
}

This was already done in the feature/rdfauthor-formbuilder branch.